feat: implement job creation and provider browsing endpoints
- Add POST /api/jobs endpoint for job creation with photo upload
  * Support multipart form data with JSON and image files
  * Validate budget constraints, category existence, and file limits
  * Handle enum types for Location and Urgency properly
  * Store job photos with proper entity relationships

- Add GET /api/providers endpoint for provider browsing
  * Support filtering by category, location, rating, and verification status
  * Implement pagination with configurable sorting options
  * Use JPA Specifications for dynamic query building
  * Return structured DTOs with provider profile information

- Fix entity validation annotations (@NotNull for enums instead of @NotBlank)
- Update security permissions to allow HYBRID_PROVIDER role for job posting
- Integrate file storage service with proper error handling

Developer: Nizar
Reviewed by: Israa

+++ backend/src/main/java/com/elwaseet/backend/dto/ProviderResponseDTO.java
@@ -1,6 +1,6 @@
 package com.elwaseet.backend.dto;
 
-
+import java.math.BigDecimal;
 import com.elwaseet.backend.entity.ProviderProfile;
 import lombok.Data;
 
@@ -12,10 +12,10 @@ public class ProviderResponseDTO {
     private String serviceAreas;
     private String availabilityDescription;
     private Boolean isVerified;
-    private Double averageRating;
+    private BigDecimal averageRating;
     private Integer totalReviews;
     private Integer totalJobsCompleted;
-    private Double totalEarned;
+    private BigDecimal totalEarned;
 
     public static ProviderResponseDTO fromEntity(ProviderProfile provider) {
         ProviderResponseDTO dto = new ProviderResponseDTO();
@@ -32,13 +32,13 @@ public class ProviderResponseDTO {
         dto.setIsVerified(provider.getIsVerified());
 
         if (provider.getAverageRating() != null) {
-            dto.setAverageRating(provider.getAverageRating().doubleValue());
+            dto.setAverageRating(provider.getAverageRating());
         }
         dto.setTotalReviews(provider.getTotalReviews());
         dto.setTotalJobsCompleted(provider.getTotalJobsCompleted());
 
         if (provider.getTotalEarned() != null) {
-            dto.setTotalEarned(provider.getTotalEarned().doubleValue());
+            dto.setTotalEarned(provider.getTotalEarned());
         }
 
         return dto;
